Fix CI: extract RecoveryProgress and DegradedDcMonitor, add ControllerRoute test entry
- Extract RecoveryProgress and StoreVersionPair into top-level
RecoveryProgress.java
  to keep DegradedModeRecoveryService under 500-line enforce-lines-added
limit.
- Extract monitor logic (duration metrics, orphan detection) into
DegradedDcMonitor.java.
- Add GET_RECOVERY_PROGRESS to ControllerRouteDimensionTest expected values.
