+++ services/venice-controller/src/main/java/com/linkedin/venice/controller/VeniceParentHelixAdmin.java
@@ -3089,7 +3089,7 @@ public class VeniceParentHelixAdmin implements Admin {
   }
 
   @Override
-  public DegradedModeRecoveryService.RecoveryProgress getRecoveryProgress(String clusterName, String datacenterName) {
+  public RecoveryProgress getRecoveryProgress(String clusterName, String datacenterName) {
     return degradedModeRecoveryService.getRecoveryProgress(clusterName, datacenterName);
   }
